Support multiple sheets in the obvious way:
New option "-i", signifies input file.

Change "-l" and "-c" to indicate the type of the input (label lines,
or code).  Let absence of delimiter signify that this is one label, to
be mapped over the whole sheet.  Don't ever fill up remainder portion
when doing blank labels -- from now on, it's either one label to be
iterated, or a bunch of labels each to be printed once.

Change usage message accordingly.

Added/changed tests for above.


git-svn-id: http://svn.red-bean.com/repos/labelnation/trunk@47 d37bbd87-4cf8-0310-9f5b-e7918974f4aa